+++ tests/integration/test_bottle.py
@@ -1,6 +1,7 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import datetime as dt
 import sys
 from contextlib import contextmanager
 
@@ -9,8 +10,12 @@ from webtest import TestApp
 
 from scout_apm.api import Config
 from scout_apm.bottle import ScoutPlugin
+from scout_apm.compat import datetime_to_timestamp
 from tests.compat import mock
-from tests.integration.util import parametrize_user_ip_headers
+from tests.integration.util import (
+    parametrize_queue_time_header_name,
+    parametrize_user_ip_headers,
+)
 
 
 @contextmanager
@@ -99,6 +104,44 @@ def test_user_ip_error(tracked_requests):
 
     tracked_request = tracked_requests[0]
     assert "user_ip" not in tracked_request.tags
+
+
+@parametrize_queue_time_header_name
+def test_queue_time(header_name, tracked_requests):
+    # Not testing floats due to Python 2/3 rounding differences
+    queue_start = int(datetime_to_timestamp(dt.datetime.utcnow()) - 2)
+    with app_with_scout() as app:
+        response = TestApp(app).get(
+            "/", headers={header_name: str("t=") + str(queue_start)}
+        )
+
+    assert response.status_int == 200
+    assert len(tracked_requests) == 1
+    queue_time_ns = tracked_requests[0].tags["scout.queue_time_ns"]
+    # Upper bound assumes we didn't take more than 2s to run this test...
+    assert queue_time_ns >= 2000000000 and queue_time_ns < 4000000000
+
+
+def test_queue_time_error(tracked_requests):
+    """
+    Scout doesn't crash if bottle's Request.headers.get raises an exception.
+
+    This cannot be tested without mocking because it should never happen.
+    """
+    orig_headers_get = WSGIHeaderDict.get
+
+    def crashy_get(self, key, *args, **kwargs):
+        if key == "x-queue-start":
+            raise ValueError("Don't try get *that* header!")
+        return orig_headers_get(key, *args, **kwargs)
+
+    with mock.patch.object(
+        WSGIHeaderDict, "get", new=crashy_get
+    ), app_with_scout() as app:
+        TestApp(app).get("/")
+
+    tracked_request = tracked_requests[0]
+    assert "scout.queue_time_ns" not in tracked_request.tags
 
 
 def test_home_ignored(tracked_requests):